new features

+++ dir_app/views.py
@@ -26,8 +26,9 @@ def home(request):
     
     topics=Topic.objects.all()
     room_count = rooms.count()
+    room_messages= Message.objects.filter(Q(room__topic__name__icontains=q))
 
-    content_dict={'rooms':rooms, 'topics':topics, 'room_count':room_count}
+    content_dict={'rooms':rooms, 'topics':topics, 'room_count':room_count, 'room_messages':room_messages}
     return render(request, 'dir_app/home.html', content_dict) 
 
 def LoginPage(request):
@@ -73,16 +74,23 @@ def registerPage(request):
 def room(request,pk):
     room = Room.objects.get(id=pk)
     room_messages =room.message_set.all().order_by('-create')
+    participants = room.participants.all()
     if request.method == 'POST':
         message = Message.objects.create(
             user=request.user,
             room=room,
             body=request.POST.get('body')
         )
+        room.participants.add(request.user)
         return redirect('room', pk=room.id)
 
-    content_dict={'room':room, 'room_messages':room_messages}
+    content_dict={'room':room, 'room_messages':room_messages,"participants":participants}
     return render(request, 'dir_app/room.html',content_dict)
+
+def userProfile(request,pk):
+    user=User.objects.get(id=pk)
+    content_dict={'user':user}
+    return render(request, 'dir_app/profile.html',content_dict)
 
 @login_required(login_url='login')
 def createRoom(request):
@@ -116,10 +124,22 @@ def updateRoom(request, pk):
 @login_required(login_url='login')
 def deleteRoom(request,pk):
     room =Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse('You not allowed here')
     if request.method == 'POST':
         room.delete()
         return redirect('home')
     return render(request, 'dir_app/delete.html', {'obj':room})
+
+@login_required(login_url='login')
+def deleteMessage(request,pk):
+    message =Message.objects.get(id=pk)
+    if request.user != message.user:
+        return HttpResponse('You not allowed here')
+    if request.method == 'POST':
+        message.delete()
+        return redirect('home')
+    return render(request, 'dir_app/delete.html', {'obj':message})
 
 # Create your views here.
